+++ packages/video-file-receiver/src/index.js
@@ -1,0 +1,1 @@
+export { VideoFileReceiverAppliance } from './VideoFileReceiverAppliance'