+++ packages/video-http-receiver/src/index.js
@@ -1,0 +1,1 @@
+export { VideoHttpReceiverAppliance } from './VideoHttpReceiverAppliance'